#3169 Split off the AdminToolsController into many different controllers for better organization.
