+++ app/Http/Controllers/AdminToolsController.php
@@ -4,67 +4,22 @@
 
 namespace App\Http\Controllers;
 
-use App\Exceptions\Logging\HandlerLoggingInterface;
-use App\Http\Controllers\Traits\ChangesMapping;
-use App\Jobs\RefreshEnemyForces;
-use App\Jobs\RegenerateCombatLogRoute;
-use App\Logic\MDT\Data\MDTDungeon;
-use App\Logic\MDT\Exception\ImportWarning;
-use App\Logic\MDT\Exception\InvalidMDTStringException;
-use App\Models\CombatLog\ChallengeModeRun;
-use App\Models\Dungeon;
 use App\Models\DungeonRoute\DungeonRoute;
 use App\Models\Floor\Floor;
-use App\Models\GameVersion\GameVersion;
-use App\Models\Mapping\MappingVersion;
-use App\Models\MDTImport;
-use App\Models\Npc\Npc;
-use App\Models\Npc\NpcClassification;
-use App\Models\Npc\NpcDungeon;
-use App\Models\Npc\NpcEnemyForces;
-use App\Models\Npc\NpcType;
-use App\Models\Spell\Spell;
 use App\Models\User;
-use App\Repositories\Interfaces\SpellRepositoryInterface;
 use App\Service\Cache\CacheServiceInterface;
 use App\Service\CombatLog\ResultEventDungeonRouteServiceInterface;
-use App\Service\Coordinates\CoordinatesServiceInterface;
-use App\Service\DungeonRoute\ThumbnailService;
 use App\Service\MapContext\MapContextServiceInterface;
-use App\Service\MDT\MDTExportStringServiceInterface;
-use App\Service\MDT\MDTImportStringServiceInterface;
-use App\Service\MDT\MDTMappingExportServiceInterface;
-use App\Service\MDT\MDTMappingImportServiceInterface;
-use App\Service\MDT\MDTMappingVersionServiceInterface;
-use App\Service\MessageBanner\MessageBannerServiceInterface;
 use App\Service\ReadOnlyMode\ReadOnlyModeServiceInterface;
-use App\Traits\SavesArrayToJsonFile;
 use Artisan;
 use Exception;
-use HaydenPierce\ClassFinder\ClassFinder;
-use Illuminate\Contracts\Foundation\Application;
-use Illuminate\Contracts\View\Factory;
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Illuminate\Session\TokenMismatchException;
-use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\Config;
-use Illuminate\Support\Facades\Log;
 use Illuminate\View\View;
-use Laravel\Pennant\Feature;
 use Session;
-use Throwable;
 
 class AdminToolsController extends Controller
 {
-    use ChangesMapping;
-    use SavesArrayToJsonFile;
-
-    /**
-     * @return Factory|View
-     */
     public function index(): View
     {
         return view('admin.tools.list');
@@ -121,870 +76,6 @@ class AdminToolsController extends Controller
         ]);
     }
 
-    /**
-     * @return Application|Factory|View
-     */
-    public function messageBanner(): View
-    {
-        return view('admin.tools.messagebanner.set');
-    }
-
-    /**
-     * @return RedirectResponse
-     */
-    public function messageBannerSubmit(
-        Request                       $request,
-        MessageBannerServiceInterface $messageBannerService,
-    ): RedirectResponse {
-        $message = $request->get('message');
-        $messageBannerService->setMessage(empty($message) ? null : $message);
-
-        Session::flash('status', __('controller.admintools.flash.message_banner_set_successfully'));
-
-        return redirect()->route('admin.tools.messagebanner.set');
-    }
-
-    /**
-     * @return Application|Factory|View
-     */
-    public function npcimport(): View
-    {
-        return view('admin.tools.npc.import');
-    }
-
-    /**
-     * @return void
-     */
-    public function npcimportsubmit(Request $request): void
-    {
-        $importString = $request->get('import_string');
-
-        // Correct the string since wowhead sucks
-        $importString = str_replace('[Listview.extraCols.popularity]', '["Listview.extraCols.popularity"]', (string)$importString);
-
-        $decoded = json_decode($importString, true);
-
-        $log = [];
-
-        // Wowhead type => keystone.guru type
-        $npcTypeMapping = [
-            15 => NpcType::ABERRATION,
-            1  => NpcType::BEAST,
-            8  => NpcType::CRITTER,
-            3  => NpcType::DEMON,
-            2  => NpcType::DRAGONKIN,
-            4  => NpcType::ELEMENTAL,
-            5  => NpcType::GIANT,
-            7  => NpcType::HUMANOID,
-            9  => NpcType::MECHANICAL,
-            6  => NpcType::UNDEAD,
-            10 => NpcType::UNCATEGORIZED,
-            // 12 => NpcType::BATTLE_PET,
-        ];
-
-        $aggressivenessMapping = [
-            -1 => 'aggressive',
-            0  => 'neutral',
-            1  => 'friendly',
-        ];
-
-        $classificationMapping = [
-            0 => NpcClassification::ALL[NpcClassification::NPC_CLASSIFICATION_NORMAL],
-            1 => NpcClassification::ALL[NpcClassification::NPC_CLASSIFICATION_NORMAL],
-            2 => NpcClassification::ALL[NpcClassification::NPC_CLASSIFICATION_ELITE],
-            3 => NpcClassification::ALL[NpcClassification::NPC_CLASSIFICATION_BOSS],
-            4 => NpcClassification::ALL[NpcClassification::NPC_CLASSIFICATION_RARE],
-        ];
-
-        try {
-            foreach ($decoded['data'] as $npcData) {
-                $npcCandidate = Npc::findOrNew($npcData['id']);
-
-                $beforeModel = clone $npcCandidate;
-
-                /** @var Dungeon $dungeon */
-                $dungeons = Dungeon::whereIn('zone_id', $npcData['location'])->get();
-                if ($dungeons->isEmpty()) {
-                    $log[] = sprintf('*** Unable to find dungeon(s) with zone_id(s) %s; npc %s (%s) NOT added; needs manual work', implode(', ', $npcData['location']), $npcData['id'], $npcData['name']);
-
-                    continue;
-                }
-
-                if (!isset($npcTypeMapping[$npcData['type']])) {
-                    $log[] = sprintf('*** Unable to find npc type %s; npc %s (%s) NOT added; needs manual work', $npcData['type'], $npcData['id'], $npcData['name']);
-
-                    continue;
-                }
-
-                $npcCandidate->id                = $npcData['id'];
-                $npcCandidate->classification_id = $classificationMapping[($npcData['classification'] ?? 0) + ($npcData['boss'] ?? 0) + 1];
-                // Bosses
-                if ($npcCandidate->classification_id >= NpcClassification::ALL[NpcClassification::NPC_CLASSIFICATION_BOSS]) {
-                    $npcCandidate->dangerous = true;
-                }
-
-                $npcCandidate->npc_type_id = $npcTypeMapping[$npcData['type']];
-                // This will be converted to a translation with localization:exportnpcnames!
-                $npcCandidate->name = $npcData['name'];
-
-                $npcCandidate->aggressiveness = isset($npcData['react']) && is_array($npcData['react']) ? $aggressivenessMapping[$npcData['react'][0] ?? -1] : 'aggressive';
-
-                $existed = $npcCandidate->exists;
-                if ($npcCandidate->save()) {
-                    foreach ($dungeons as $dungeon) {
-                        NpcDungeon::create([
-                            'npc_id'     => $npcCandidate->id,
-                            'dungeon_id' => $dungeon->id,
-                        ]);
-
-                        if ($existed) {
-                            $log[] = sprintf('Updated NPC %s (%s) in %s', $npcData['id'], $npcData['name'], __($dungeon->name));
-                        } else {
-                            // Now create new enemy forces. Default to 0
-                            $npcCandidate->createNpcEnemyForcesForExistingMappingVersions();
-
-                            $log[] = sprintf('Inserted NPC %s (%s) in %s', $npcData['id'], $npcData['name'], __($dungeon->name));
-                        }
-                    }
-                } else {
-                    $log[] = sprintf('*** Error saving NPC %s (%s)', $npcData['id'], $npcData['name']);
-                }
-
-                // Changed the mapping; so make sure we synchronize it
-                $this->mappingChanged($beforeModel, $npcCandidate);
-            }
-
-            // Cannot do this automatically due to permission issues
-            $log[] = '';
-            $log[] = 'You can now run php artisan localization:exportnpcnames to export the NPC names to the localization files.';
-            $log[] = 'You can now run php artisan localization:importnpcnames to convert the NPC names to translation keys in the database.';
-        } catch (Exception $exception) {
-            dump($exception);
-        } finally {
-            dump($log);
-        }
-    }
-
-    /**
-     * @param  Dungeon|null $dungeon
-     * @return View
-     */
-    public function manageSpellVisibility(Request $request, ?Dungeon $dungeon = null): View
-    {
-        return view('admin.tools.npc.managespellvisibility', [
-            'npcs' => Npc::when($dungeon !== null, fn(Builder $builder) => $builder->join('npc_dungeons', 'npc_dungeons.npc_id', '=', 'npcs.id')
-                ->select('npcs.*')
-                ->where('npc_dungeons.dungeon_id', $dungeon->id))->with('npcSpells')
-                ->has('npcSpells')
-                ->paginate(50),
-            'spells' => Spell::with('gameVersion')->when($dungeon !== null, fn(Builder $builder) => $builder->whereRelation('spellDungeons', 'dungeon_id', $dungeon->id))->get()
-                ->keyBy('id'),
-            'dungeon' => $dungeon,
-        ]);
-    }
-
-    public function manageSpellVisibilitySubmit(Request $request): RedirectResponse
-    {
-        $dungeonId = (int)$request->get('dungeon_id');
-        $dungeon   = null;
-        if ($dungeonId !== -1) {
-            $dungeon = Dungeon::findOrFail($dungeonId);
-        }
-
-        return redirect()->route('admin.tools.npc.managespellvisibility', ['dungeon' => $dungeon]);
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function dungeonroute(): View
-    {
-        return view('admin.tools.dungeonroute.view');
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function dungeonroutesubmit(Request $request): View
-    {
-        $publicKey = $request->get('public_key');
-
-        $dungeonRoute = DungeonRoute::with([
-            'faction',
-            'specializations',
-            'classes',
-            'races',
-            'affixes',
-            'brushlines',
-            'paths',
-            'author',
-            'killZones',
-            'pridefulEnemies',
-            'publishedstate',
-            'ratings',
-            'favorites',
-            'enemyraidmarkers',
-            'mapicons',
-            'mdtImport',
-            'team',
-        ])->when(is_numeric($publicKey), function (Builder $builder) use ($publicKey) {
-            $builder->where('id', $publicKey);
-        }, function (Builder $builder) use ($publicKey) {
-            $builder->where('public_key', $publicKey);
-        })->firstOrFail();
-
-        return view('admin.tools.dungeonroute.viewcontents', [
-            'dungeonroute' => $dungeonRoute,
-        ]);
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function dungeonrouteMappingVersions(): View
-    {
-        $mappingVersionUsage = MappingVersion::orderBy('dungeon_id')
-            ->get()
-            ->mapWithKeys(static fn(
-                MappingVersion $mappingVersion,
-            ) => [$mappingVersion->getPrettyName() => $mappingVersion->dungeonRoutes()->count()])
-            ->groupBy(static fn(int $count, string $key) => $count === 0, true);
-
-        return view('admin.tools.dungeonroute.mappingversions', [
-            'mappingVersionUsage' => collect([
-                'unused' => $mappingVersionUsage[1],
-                'used'   => $mappingVersionUsage[0],
-            ]),
-        ]);
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function enemyforcesimport(): View
-    {
-        return view('admin.tools.enemyforces.import');
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function enemyforcesimportsubmit(Request $request): void
-    {
-        $json = json_decode((string)$request->get('import_string'), true);
-
-        $results = [];
-        foreach ($json['Npcs'] as $jsonNpc) {
-            $npc = Npc::where('id', $jsonNpc['Id'])->first();
-
-            if ($npc !== null) {
-                $keyMapping = [
-                    'MythicHealth' => 'base_health',
-                    'Amount'       => 'enemy_forces',
-                ];
-
-                $toUpdate = [];
-                foreach ($jsonNpc as $key => $value) {
-                    if ($key !== 'Id' && $value >= 0 && isset($keyMapping[$key])) {
-                        $toUpdate[$keyMapping[$key]] = $value;
-                    }
-                }
-
-                $npc->update($toUpdate);
-
-                $results[] = sprintf('Changed npc %d fields: %s', $jsonNpc['Id'], json_encode($toUpdate));
-            } else {
-                $results[] = sprintf('Unable to find npc %d', $jsonNpc['Id']);
-            }
-        }
-
-        dd($results);
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function enemyforcesrecalculate(): View
-    {
-        return view('admin.tools.enemyforces.recalculate');
-    }
-
-    /**
-     * @return void
-     */
-    public function enemyforcesrecalculatesubmit(Request $request): void
-    {
-        $dungeonId = (int)$request->get('dungeon_id');
-
-        $builder = DungeonRoute::without([
-            'faction',
-            'specializations',
-            'classes',
-            'races',
-            'affixes',
-        ])
-            ->select('id')
-            ->when($dungeonId !== -1, static fn(Builder $builder) => $builder->where('dungeon_id', $dungeonId));
-
-        $count = 0;
-        foreach ($builder->get() as $dungeonRoute) {
-            RefreshEnemyForces::dispatch($dungeonRoute->id);
-            $count++;
-        }
-
-        dd(sprintf('Dispatched %d jobs', $count));
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function thumbnailsregenerate(): View
-    {
-        return view('admin.tools.thumbnails.regenerate');
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function thumbnailsregeneratesubmit(Request $request, ThumbnailService $thumbnailService): View
-    {
-        set_time_limit(3600);
-
-        $dungeonId   = (int)$request->get('dungeon_id');
-        $onlyMissing = (int)$request->get('only_missing');
-
-        $builder = DungeonRoute::without([
-            'faction',
-            'specializations',
-            'classes',
-            'races',
-            'affixes',
-        ])
-            ->with('dungeon')
-            ->when($dungeonId !== -1, static fn(Builder $builder) => $builder->where('dungeon_id', $dungeonId))
-            ->orderByDesc('created_at');
-
-        $successCount  = 0;
-        $failureCount  = 0;
-        $dungeonRoutes = $builder->get();
-        foreach ($dungeonRoutes as $dungeonRoute) {
-            $shouldRefresh = !$onlyMissing || !$thumbnailService->hasThumbnailsGenerated($dungeonRoute);
-
-            if ($shouldRefresh) {
-                if ($thumbnailService->queueThumbnailRefresh($dungeonRoute)) {
-                    $successCount++;
-                } else {
-                    $failureCount++;
-                }
-            }
-        }
-
-        Session::flash('status', __('controller.admintools.flash.thumbnail_regenerate_result', [
-            'success' => $successCount,
-            'total'   => $successCount + $failureCount,
-            'failed'  => $failureCount,
-        ]));
-
-        return view('admin.tools.thumbnails.regenerate');
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function combatlogregenerate(): View
-    {
-        return view('admin.tools.combatlog.regenerate');
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function combatlogregeneratesubmit(Request $request): View
-    {
-        set_time_limit(3600);
-
-        $dungeonId = (int)$request->get('dungeon_id');
-
-        $count = 0;
-
-        // Cannot use joins since the other table lives in a different database
-        DungeonRoute::query()
-            ->when($dungeonId !== -1, static fn(Builder $builder) => $builder->where('dungeon_id', $dungeonId))
-            ->chunkById(200, function (Collection $dungeonRoutes) use (&$count) {
-                $dungeonRoutes = $dungeonRoutes->keyBy('id');
-                /** @var Collection<ChallengeModeRun> $challengeModes */
-                $challengeModes = ChallengeModeRun::whereIn('dungeon_route_id', $dungeonRoutes->pluck('id'))
-                    ->get();
-
-                foreach ($challengeModes as $challengeMode) {
-                    RegenerateCombatLogRoute::dispatch(
-                        $dungeonRoutes->get($challengeMode->dungeon_route_id)->id,
-                    );
-                    $count++;
-                }
-            });
-
-        Session::flash('status', __('controller.admintools.flash.combatlog_route_regenerate_result', [
-            'count' => $count,
-        ]));
-
-        return view('admin.tools.combatlog.regenerate');
-    }
-
-    /**
-     * @return Factory|
-     */
-    public function mdtview(): View
-    {
-        return view('admin.tools.mdt.string');
-    }
-
-    public function mdtviewsubmit(
-        Request                         $request,
-        MDTImportStringServiceInterface $mdtImportStringService,
-    ): JsonResponse {
-        return response()->json(
-            $mdtImportStringService
-                ->setEncodedString($request->get('import_string'))
-                ->getDecoded(),
-        );
-    }
-
-    /**
-     * @return Factory|
-     */
-    public function mdtviewasdungeonroute(): View
-    {
-        return view('admin.tools.mdt.string', ['asDungeonroute' => true]);
-    }
-
-    public function mdtImportList(): View
-    {
-        return view('admin.tools.mdt.list', [
-            'mdtImports' => MDTImport::whereNotNull('error')
-                ->orderByDesc('created_at')
-                ->paginate(50),
-        ]);
-    }
-
-    /**
-     * @return never|void
-     *
-     * @throws Throwable
-     */
-    public function mdtviewasdungeonroutesubmit(
-        Request                         $request,
-        MDTImportStringServiceInterface $mdtImportStringService,
-    ) {
-        try {
-            $dungeonRoute = $mdtImportStringService
-                ->setEncodedString($request->get('import_string'))
-                ->getDungeonRoute(collect(), collect());
-            $dungeonRoute->makeVisible([
-                'affixes',
-                'killZones',
-            ]);
-
-            dd($dungeonRoute);
-        } catch (InvalidMDTStringException) {
-            return abort(400, __('controller.admintools.error.mdt_string_format_not_recognized'));
-        } catch (Exception $ex) {
-            // Different message based on our deployment settings
-            if (config('app.debug')) {
-                $message = sprintf(__('controller.admintools.error.invalid_mdt_string_exception'), $ex->getMessage());
-            } else {
-                $message = __('controller.admintools.error.invalid_mdt_string');
-            }
-
-            return abort(400, $message);
-        } catch (Throwable $error) {
-            if ($error->getMessage() === "Class 'Lua' not found") {
-                return abort(500, __('controller.admintools.error.mdt_importer_not_configured'));
-            }
-
-            throw $error;
-        }
-    }
-
-    /**
-     * @return Factory|
-     */
-    public function mdtviewasstring(): View
-    {
-        return view('admin.tools.mdt.dungeonroute');
-    }
-
-    /**
-     * @return never|void
-     *
-     * @throws Throwable
-     */
-    public function mdtviewasstringsubmit(
-        Request                         $request,
-        MDTImportStringServiceInterface $mdtImportStringService,
-        MDTExportStringServiceInterface $mdtExportStringService,
-    ) {
-        $publicKey = $request->get('public_key');
-
-        $dungeonRoute = DungeonRoute::when(is_numeric($publicKey), function (Builder $builder) use ($publicKey) {
-            $builder->where('id', $publicKey);
-        }, function (Builder $builder) use ($publicKey) {
-            $builder->where('public_key', $publicKey);
-        })->firstOrFail();
-
-        try {
-            $warnings = new Collection();
-
-            $exportString = $mdtExportStringService
-                ->setDungeonRoute($dungeonRoute)
-                ->getEncodedString($warnings);
-
-            $stringContents = $mdtImportStringService
-                ->setEncodedString($exportString)
-                ->getDecoded();
-
-            dd($exportString, $stringContents);
-        } catch (Exception $ex) {
-            // Different message based on our deployment settings
-            if (config('app.debug')) {
-                $message = sprintf(__('controller.admintools.error.invalid_mdt_string_exception'), $ex->getMessage());
-            } else {
-                $message = __('controller.admintools.error.invalid_mdt_string');
-            }
-
-            return abort(400, $message);
-        } catch (Throwable $error) {
-            if ($error->getMessage() === "Class 'Lua' not found") {
-                return abort(500, __('controller.admintools.error.mdt_importer_not_configured'));
-            }
-
-            throw $error;
-        }
-    }
-
-    /**
-     * @return Factory|
-     */
-    public function mdtdungeonmappinghash(): View
-    {
-        return view('admin.tools.mdt.dungeonmappinghash');
-    }
-
-    /**
-     * @return void
-     *
-     * @throws Throwable
-     */
-    public function mdtdungeonmappinghashsubmit(
-        Request                          $request,
-        MDTMappingImportServiceInterface $mdtMappingService,
-    ): void {
-        $dungeon = Dungeon::findOrFail($request->get('dungeon_id'));
-
-        dd($mdtMappingService->getMDTMappingHash($dungeon->key));
-    }
-
-    /**
-     * @return Factory|
-     */
-    public function dungeonmappingversiontomdtmapping()
-    {
-        return view('admin.tools.mdt.dungeonmappingversiontomdtmapping', [
-            'mappingVersionsSelect' => MappingVersion::orderBy('dungeon_id')
-                ->get()
-                ->groupBy('dungeon_id')
-                ->mapWithKeys(static function (Collection $mappingVersionByDungeon, int $id) {
-                    $dungeon = Dungeon::findOrFail($id);
-
-                    return [
-                        __($dungeon->name) => $mappingVersionByDungeon->mapWithKeys(static fn(
-                            MappingVersion $mappingVersion,
-                        ) => [
-                            $mappingVersion->id => $mappingVersion->getPrettyName(),
-                        ]),
-                    ];
-                }),
-        ]);
-    }
-
-    /**
-     * @return void
-     *
-     * @throws Throwable
-     */
-    public function dungeonmappingversiontomdtmappingsubmit(
-        Request                          $request,
-        MDTMappingExportServiceInterface $mdtMappingService,
-    ): void {
-        $mappingVersion = MappingVersion::findOrFail($request->get('mapping_version_id'));
-
-        echo $mdtMappingService->getMDTMappingAsLuaString($mappingVersion);
-        dd();
-    }
-
-    public function dungeonMappingVersionAccuracy(
-        Request                           $request,
-        MDTMappingVersionServiceInterface $mappingVersionService,
-    ): View {
-        /** @var Collection<Dungeon> $allDungeons */
-        $allDungeons = Dungeon::with(['mappingVersions', 'mappingVersions.dungeon'])->get();
-
-        $dungeonAccuracyByFloor = collect();
-        foreach ($allDungeons as $dungeon) {
-            /**
-             * @var MappingVersion $latestMappingVersion I load the latest mapping version for this dungeon regardless
-             *                     of game version because we want to compare to the latest MDT version. The newest mapping version will
-             *                     generally be the mapping version that was used in the most recent version of MDT.
-             */
-            $latestMappingVersion = $dungeon->mappingVersions->first();
-
-            $dungeonAccuracyByFloor->put(
-                $dungeon->id,
-                $mappingVersionService->getMappingVersionAccuracy(
-                    $latestMappingVersion,
-                ),
-            );
-        }
-
-        return view('admin.tools.mdt.dungeonmappingversionaccuracy', [
-            'dungeonAccuracyByFloor' => $dungeonAccuracyByFloor,
-            'dungeonsById'           => $allDungeons->keyBy('id'),
-        ]);
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function spellsShowMissingSpellInfo(
-        SpellRepositoryInterface $spellRepository,
-    ): View {
-        $missingSpells = $spellRepository->getMissingSpellIds();
-
-        return view('admin.tools.spells.showmissingspellinfo', [
-            'spells' => Spell::whereNull('fetched_data_at')->get()->merge(
-                collect($missingSpells)->map(fn($spellId) => new Spell(['id' => $spellId, 'name' => 'Unknown'])),
-            ),
-        ]);
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function npcsShowMissingDisplayId(): View
-    {
-        return view('admin.tools.npc.showmissingdisplayid', [
-            'npcs' => Npc::whereNull('display_id')->get(),
-        ]);
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function wagoggImportIngameCoordinates(): View
-    {
-        return view('admin.tools.wagogg.importingamecoordinates');
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     *
-     * @throws Exception
-     */
-    public function wagoggImportIngameCoordinatesSubmit(Request $request): void
-    {
-        // Parse all UI Map Assignment TABLE data and convert them to a workable format
-        $uiMapAssignmentTable                   = $request->get('ui_map_assignment_table_csv');
-        $uiMapAssignmentTableParsed             = str_getcsv_assoc($uiMapAssignmentTable);
-        $uiMapAssignmentTableHeaders            = array_shift($uiMapAssignmentTableParsed);
-        $uiMapAssignmentTableHeaderIndexUiMapId = array_search('UiMapID', $uiMapAssignmentTableHeaders, true);
-        $uiMapAssignmentTableHeaderIndexMinX    = array_search('Region_0', $uiMapAssignmentTableHeaders, true);
-        $uiMapAssignmentTableHeaderIndexMinY    = array_search('Region_1', $uiMapAssignmentTableHeaders, true);
-        $uiMapAssignmentTableHeaderIndexMaxX    = array_search('Region_3', $uiMapAssignmentTableHeaders, true);
-        $uiMapAssignmentTableHeaderIndexMaxY    = array_search('Region_4', $uiMapAssignmentTableHeaders, true);
-
-        /** @var Collection<Floor> $allFloors */
-        //        $allDungeons = Dungeon::where('key', Dungeon::DUNGEON_AZJOL_NERUB)->get()->keyBy('id');
-        $allFloors = Floor::where('facade', 0)
-            ->where('ui_map_id', '>', 0)
-            ->where('ingame_min_x', 0)
-            ->where('ingame_min_y', 0)
-//            ->where('ingame_max_x', 0)
-//            ->where('ingame_max_y', 0)
-            ->get();
-
-        dump('Changed floors:');
-
-        $allUiMapIds                = $allFloors->pluck('ui_map_id')->toArray();
-        $uiMapAssignmentTableParsed = array_filter($uiMapAssignmentTableParsed, fn(array $item) => in_array($item[$uiMapAssignmentTableHeaderIndexUiMapId], $allUiMapIds));
-
-        // Go over the UI Map Assignments and find the ones we're interested in
-        foreach ($allFloors as $floor) {
-            foreach ($uiMapAssignmentTableParsed as $index => $uiMapAssignmentRow) {
-                if (((int)$uiMapAssignmentRow[$uiMapAssignmentTableHeaderIndexUiMapId]) === $floor->ui_map_id) {
-                    $beforeModel = clone $floor;
-
-                    $floor->update([
-                        'ingame_min_x' => round((float)$uiMapAssignmentRow[$uiMapAssignmentTableHeaderIndexMinX], 2),
-                        'ingame_min_y' => round($uiMapAssignmentRow[$uiMapAssignmentTableHeaderIndexMinY], 2),
-                        'ingame_max_x' => round($uiMapAssignmentRow[$uiMapAssignmentTableHeaderIndexMaxX], 2),
-                        'ingame_max_y' => round($uiMapAssignmentRow[$uiMapAssignmentTableHeaderIndexMaxY], 2),
-                    ]);
-
-                    dump(sprintf('Updated floor %s (id: %d, ui_map_id: %d) ', __($floor->name), $floor->id, $floor->ui_map_id));
-
-//                    $this->mappingChanged($beforeModel, $floor);
-
-                    break;
-                }
-            }
-        }
-
-        dd('done!');
-    }
-
-    /**
-     * @return Factory|
-     *
-     * @throws Exception
-     */
-    public function mdtdiff(
-        CacheServiceInterface       $cacheService,
-        CoordinatesServiceInterface $coordinatesService,
-    ): View {
-        $warnings = new Collection();
-        $npcs     = Npc::with([
-            'npcEnemyForces',
-            'enemies',
-            'type',
-        ])->get();
-
-        // For each dungeon
-        foreach (Dungeon::all() as $dungeon) {
-            /** @var Dungeon $dungeon */
-            $mdtNpcs = new MDTDungeon($cacheService, $coordinatesService, $dungeon)->getMDTNPCs();
-
-            // For each NPC that is found in the MDT Dungeon
-            foreach ($mdtNpcs as $mdtNpc) {
-                // Ignore mobs we should ignore
-                if (!$mdtNpc->isValid() || $mdtNpc->isAwakened()) {
-                    continue;
-                }
-
-                // Find our own NPC
-                /** @var Npc $npc */
-                $npc = $npcs->where('id', $mdtNpc->getId())->first();
-
-                // Not found..
-                if ($npc === null) {
-                    $warnings->push(
-                        new ImportWarning(
-                            'missing_npc',
-                            sprintf(__('controller.admintools.error.mdt_unable_to_find_npc_for_id'), $mdtNpc->getId()),
-                            [
-                                'mdt_npc' => (object)$mdtNpc->getRawMdtNpc(),
-                                'npc'     => $npc,
-                            ],
-                        ),
-                    );
-                } // Found, compare
-                else {
-                    // Match health
-                    $npcHealth = $npc->getHealthByGameVersion(GameVersion::firstWhere('key', GameVersion::GAME_VERSION_RETAIL));
-                    if ($npcHealth?->health !== $mdtNpc->getHealth()) {
-                        $warnings->push(
-                            new ImportWarning(
-                                'mismatched_health',
-                                sprintf(__('controller.admintools.error.mdt_mismatched_health'), $mdtNpc->getId(), $mdtNpc->getHealth(), $npcHealth->health),
-                                [
-                                    'mdt_npc' => (object)$mdtNpc->getRawMdtNpc(),
-                                    'npc'     => $npc,
-                                    'old'     => $npcHealth->health,
-                                    'new'     => $mdtNpc->getHealth(),
-                                ],
-                            ),
-                        );
-                    }
-
-                    // Match enemy forces
-                    /** @var NpcEnemyForces|null $npcEnemyForces */
-                    $npcEnemyForces = $npc->enemyForcesByMappingVersion();
-                    if ($npcEnemyForces?->enemy_forces !== $mdtNpc->getCount()) {
-                        $warnings->push(
-                            new ImportWarning(
-                                'mismatched_enemy_forces',
-                                sprintf(__('controller.admintools.error.mdt_mismatched_enemy_forces'), $mdtNpc->getId(), $mdtNpc->getCount(), $npcEnemyForces?->enemy_forces),
-                                [
-                                    'mdt_npc' => (object)$mdtNpc->getRawMdtNpc(),
-                                    'npc'     => $npc,
-                                    'old'     => $npcEnemyForces?->enemy_forces,
-                                    'new'     => $mdtNpc->getCount(),
-                                ],
-                            ),
-                        );
-                    }
-
-                    // Match enemy forces teeming
-                    if ($npcEnemyForces?->enemy_forces_teeming !== $mdtNpc->getCountTeeming()) {
-                        $warnings->push(
-                            new ImportWarning(
-                                'mismatched_enemy_forces_teeming',
-                                sprintf(__('controller.admintools.error.mdt_mismatched_enemy_forces_teeming'), $mdtNpc->getId(), $mdtNpc->getCountTeeming(), $npcEnemyForces?->enemy_forces_teeming),
-                                [
-                                    'mdt_npc' => (object)$mdtNpc->getRawMdtNpc(),
-                                    'npc'     => $npc,
-                                    'old'     => $npcEnemyForces?->enemy_forces_teeming,
-                                    'new'     => $mdtNpc->getCountTeeming(),
-                                ],
-                            ),
-                        );
-                    }
-
-                    // Match clone count, should be equal
-                    if ($npc->enemies->count() !== count($mdtNpc->getClones())) {
-                        $warnings->push(
-                            new ImportWarning(
-                                'mismatched_enemy_count',
-                                sprintf(
-                                    __('controller.admintools.error.mdt_mismatched_enemy_count'),
-                                    $mdtNpc->getId(),
-                                    count($mdtNpc->getClones()),
-                                    $npc->enemies === null ? 0 : $npc->enemies->count(),
-                                ),
-                                [
-                                    'mdt_npc' => (object)$mdtNpc->getRawMdtNpc(),
-                                    'npc'     => $npc,
-                                ],
-                            ),
-                        );
-                    }
-
-                    // Match npc type, should be equal
-                    if ($npc->type->type !== $mdtNpc->getCreatureType()) {
-                        $warnings->push(
-                            new ImportWarning(
-                                'mismatched_enemy_type',
-                                sprintf(
-                                    __('controller.admintools.error.mdt_mismatched_enemy_type'),
-                                    $mdtNpc->getId(),
-                                    $mdtNpc->getCreatureType(),
-                                    $npc->type->type,
-                                ),
-                                [
-                                    'mdt_npc' => (object)$mdtNpc->getRawMdtNpc(),
-                                    'npc'     => $npc,
-                                    'old'     => $npc->type->type,
-                                    'new'     => $mdtNpc->getCreatureType(),
-                                ],
-                            ),
-                        );
-                    }
-                }
-            }
-        }
-
-        return view('admin.tools.mdt.diff', ['warnings' => $warnings]);
-    }
-
     public function dropCache(Request $request, CacheServiceInterface $cacheService): RedirectResponse
     {
         ini_set('max_execution_time', -1);
@@ -1000,61 +91,11 @@ class AdminToolsController extends Controller
         return redirect()->route('admin.tools');
     }
 
-    /**
-     * @return void
-     */
     public function mappingForceSync(Request $request): void
     {
         Artisan::call('mapping:sync', ['--force' => true]);
     }
 
-    /**
-     * @return array
-     */
-    public function applyChange(Request $request)
-    {
-        $category  = $request->get('category');
-        $npcId     = $request->get('npc_id');
-        $dungeonId = $request->get('dungeon_id');
-        $value     = $request->get('value');
-
-        /** @var Npc $npc */
-        $npc     = Npc::with(['enemyForces'])->find($npcId);
-        $dungeon = Dungeon::findOrFail($dungeonId);
-
-        switch ($category) {
-            case 'mismatched_health':
-                $npc->getHealthByGameVersion(
-                    GameVersion::firstWhere('key', GameVersion::GAME_VERSION_RETAIL),
-                )?->update([
-                    'health' => $value,
-                ]);
-                break;
-            case 'mismatched_enemy_forces':
-                $npc->setEnemyForces($value, $dungeon->getCurrentMappingVersion());
-
-                break;
-                // Teeming is deprecated pretty much
-            //            case 'mismatched_enemy_forces_teeming':
-            //                $npc->enemyForces->enemy_forces_teeming = $value;
-            //                $npc->enemyForces->save();
-            //                break;
-            case 'mismatched_enemy_type':
-                $npcType          = NpcType::where('type', $value)->first();
-                $npc->npc_type_id = $npcType->id;
-                $npc->save();
-                break;
-            default:
-                abort(500, __('controller.admintools.error.mdt_invalid_category'));
-        }
-
-        // Whatever
-        return [];
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
     public function exportreleases(Request $request): View
     {
         Artisan::call('release:save');
@@ -1077,110 +118,5 @@ class AdminToolsController extends Controller
         }
 
         return redirect()->route('admin.tools');
-    }
-
-    /**
-     * @throws Exception
-     */
-    public function exportdungeondata(Request $request): View
-    {
-        Artisan::call('mapping:save');
-
-        return view('admin.tools.datadump.viewexporteddungeondata');
-    }
-
-    /**
-     * @return Application|Factory|\Illuminate\Contracts\View\View
-     */
-    public function exceptionselect(Request $request): View
-    {
-        return view('admin.tools.exception.select', [
-            'exceptions' => [
-                'TokenMismatchException' => 'TokenMismatchException',
-                'InternalServerError'    => 'InternalServerError',
-                'DiscordException'       => 'DiscordException',
-            ],
-        ]);
-    }
-
-    /**
-     * @throws TokenMismatchException
-     * @throws Exception
-     */
-    public function exceptionselectsubmit(Request $request): array
-    {
-        switch ($request->get('exception')) {
-            case 'TokenMismatchException':
-                throw new TokenMismatchException(__('controller.admintools.flash.exception'));
-            case 'InternalServerError':
-                throw new Exception(__('controller.admintools.flash.exception'));
-            case 'DiscordException':
-                Log::error('Manual Generic test log from web');
-
-                Log::channel('discord')->error('Manual Discord test log from web');
-
-                Log::stack([
-                    'daily',
-                    'discord',
-                ])->error('Manual stack log test from web');
-
-                $handlerLogging = app()->make(HandlerLoggingInterface::class);
-                $handlerLogging->uncaughtException(
-                    $request->ip(),
-                    $request->url(),
-                    null,
-                    null,
-                    null,
-                    'DiscordException',
-                    'Structured logging test from web',
-                );
-
-                return [
-                    'logging.default' => Config::get('logging.default'),
-                    'stack_channels'  => Config::get('logging.channels.stack.channels'),
-                    'discord_config'  => Config::get('logging.channels.discord'),
-                ];
-        }
-
-        return [];
-    }
-
-    public function listFeatures(Request $request): View
-    {
-        return view('admin.tools.features.list', [
-            'features' => collect(ClassFinder::getClassesInNamespace('App\\Features')),
-        ]);
-    }
-
-    public function toggleFeature(Request $request): RedirectResponse
-    {
-        $feature = (string)$request->get('feature');
-
-        $wasActive = Feature::active($feature);
-        if ($wasActive) {
-            Feature::deactivateForEveryone($feature);
-        } else {
-            Feature::activateForEveryone($feature);
-        }
-
-        Session::flash('status', __(!$wasActive ?
-            'controller.admintools.flash.feature_toggle_activated' :
-            'controller.admintools.flash.feature_toggle_deactivated', [
-                'feature' => $feature,
-            ]));
-
-        return redirect()->route('admin.tools.features.list');
-    }
-
-    public function forgetFeature(Request $request): RedirectResponse
-    {
-        $feature = (string)$request->get('feature');
-
-        Feature::forget($feature);
-        Feature::for(null)->forget($feature);
-
-        Session::flash('status', __('controller.admintools.flash.feature_forgotten', ['feature' => $feature]));
-
-        return redirect()->route('admin.tools.features.list');
     }
 }
