+++ lang/en_US/breadcrumbs.php
@@ -62,17 +62,31 @@ return [
         'admin'      => [
             'admin' => 'Admin',
             'tools' => [
-                'admin_tools'               => 'Admin tools',
-                'view_exported_dungeondata' => 'View exported dungeondata',
-                'view_exported_releases'    => 'View exported releases',
-                'select_exception'          => 'Select exception',
-                'mdt_diff'                  => 'MDT diff',
-                'view_mdt_string_contents'  => 'View MDT string contents',
-                'import_npcs'               => 'Import NPCs',
-                'spells_missing_info'       => 'Spells missing info',
-                'npcs_missing_display_id'   => 'NPCs missing display ID',
-                'thumbnails_regenerate'     => 'Regenerate thumbnails',
-                'combat_log_regenerate'     => 'Regenerate ARC routes',
+                'admin_tools'                                => 'Admin tools',
+                'view_exported_dungeondata'                  => 'View exported dungeondata',
+                'view_exported_releases'                     => 'View exported releases',
+                'select_exception'                           => 'Select exception',
+                'mdt_diff'                                   => 'MDT diff',
+                'view_mdt_string_contents'                   => 'View MDT string contents',
+                'import_npcs'                                => 'Import NPCs',
+                'spells_missing_info'                        => 'Spells missing info',
+                'npcs_missing_display_id'                    => 'NPCs missing display ID',
+                'thumbnails_regenerate'                      => 'Regenerate thumbnails',
+                'combat_log_regenerate'                      => 'Regenerate ARC routes',
+                'dungeonroute_view'                          => 'View dungeon route',
+                'dungeonroute_view_contents'                 => 'Route contents',
+                'dungeonroute_mapping_version_usage'         => 'Mapping version usage',
+                'enemyforces_import'                         => 'Import enemy forces',
+                'enemyforces_recalculate'                    => 'Recalculate enemy forces',
+                'features_list'                              => 'Features',
+                'mdt_dungeon_mapping_hash'                   => 'Dungeon mapping hash',
+                'mdt_dungeon_mapping_version_accuracy'       => 'Dungeon mapping version accuracy',
+                'mdt_dungeon_mapping_version_to_mdt_mapping' => 'Dungeon mapping version to MDT mapping',
+                'mdt_dungeonroute'                           => 'MDT dungeon route',
+                'mdt_list'                                   => 'MDT string list',
+                'messagebanner_set'                          => 'Set message banner',
+                'npc_manage_spell_visibility'                => 'Manage NPC spell visibility',
+                'wagogg_import_ingame_coordinates'           => 'Import ingame coordinates',
             ],
             'releases'     => 'Releases',
             'new_release'  => 'New release',
